requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
banks==2.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.3.1
dirtyjson==1.0.8
fastapi==0.128.4
fastembed==0.7.4
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.1
griffe==1.15.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
llama-index-core==0.14.13
llama-index-embeddings-fastembed==0.5.0
llama-index-embeddings-ollama==0.8.6
llama-index-instrumentation==0.4.2
llama-index-llms-ollama==0.9.1
llama-index-vector-stores-qdrant==0.9.1
llama-index-workflows==2.14.1
loguru==0.7.3
MarkupSafe==3.0.3
marshmallow==3.26.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.2
ollama==0.6.1
onnxruntime==1.23.2
packaging==26.0
pillow==11.3.0
platformdirs==4.5.1
pluggy==1.6.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.5
py_rust_stemmers==0.1.5
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
qdrant-client==1.16.2
regex==2026.1.15
requests==2.32.5
setuptools==81.0.0
shellingham==1.5.4
SQLAlchemy==2.0.46
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.3
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wrapt==2.1.1
yarl==1.22.0
